+++ vllm/model_executor/layers/quantization/compressed_tensors/schemes/compressed_tensors_w4a8_int.py
@@ -4,6 +4,8 @@
 from typing import Callable, Optional
 
 import torch
+from compressed_tensors.quantization import (QuantizationArgs,
+                                             QuantizationStrategy)
 
 from vllm.logger import init_logger
 from vllm.model_executor.layers.quantization.compressed_tensors.schemes import (
@@ -17,11 +19,28 @@ from vllm.scalar_type import scalar_types
 
 logger = init_logger(__name__)
 
-__all__ = ["CompressedTensorsW4A8Int"]
+__all__ = ["is_dynamic_token_w4a8_int", "CompressedTensorsW4A8Int"]
 W4A8_SUPPORTED_TYPES_MAP = {
     4: scalar_types.int4,
 }
 W4A8_SUPPORTED_BITS = list(W4A8_SUPPORTED_TYPES_MAP.keys())
+
+
+def is_dynamic_token_w4a8_int(weight_quant: QuantizationArgs,
+                              input_quant: Optional[QuantizationArgs]) -> bool:
+    is_weight_4_bits = weight_quant.num_bits == 4
+    is_activation_8_bits = input_quant.num_bits == 8
+    weight_strategy = (
+        weight_quant.strategy == QuantizationStrategy.GROUP.value
+        or weight_quant.strategy == QuantizationStrategy.CHANNEL.value)
+    is_token = (weight_strategy
+                and input_quant.strategy == QuantizationStrategy.TOKEN.value)
+    is_dynamic = not weight_quant.dynamic and input_quant.dynamic
+
+    # Both symmetric and asymmetric input quantization supported.
+    # Only symmetric weight quantization supported.
+    return (is_weight_4_bits and is_activation_8_bits and is_token
+            and weight_quant.symmetric and is_dynamic)
 
 
 class CompressedTensorsW4A8Int(CompressedTensorsScheme):
